Remove intermediate heap allocations when interacting with Base58 addresses
