+++ Cargo.toml
@@ -8,15 +8,16 @@ members = [
 	"bin/node/bench",
 	"bin/node/browser-testing",
 	"bin/node/cli",
-	"bin/node/test-runner-example",
 	"bin/node/executor",
 	"bin/node/primitives",
-	"bin/node/rpc",
 	"bin/node/rpc-client",
+	"bin/node/rpc",
 	"bin/node/runtime",
+	"bin/node/test-runner-example",
 	"bin/node/testing",
 	"bin/utils/chain-spec-builder",
 	"bin/utils/subkey",
+	"client/allocator",
 	"client/api",
 	"client/authority-discovery",
 	"client/basic-authorship",
@@ -43,16 +44,15 @@ members = [
 	"client/informant",
 	"client/keystore",
 	"client/light",
-	"client/network",
 	"client/network-gossip",
+	"client/network",
 	"client/network/test",
 	"client/offchain",
 	"client/peerset",
-	"client/allocator",
 	"client/proposer-metrics",
-	"client/rpc",
 	"client/rpc-api",
 	"client/rpc-servers",
+	"client/rpc",
 	"client/service",
 	"client/service/test",
 	"client/state-db",
@@ -76,13 +76,12 @@ members = [
 	"frame/contracts/rpc",
 	"frame/contracts/rpc/runtime-api",
 	"frame/democracy",
-	"frame/try-runtime",
-	"frame/elections",
 	"frame/election-provider-multi-phase",
 	"frame/election-provider-support",
-	"frame/example",
+	"frame/elections",
 	"frame/example-offchain-worker",
 	"frame/example-parallel",
+	"frame/example",
 	"frame/executive",
 	"frame/gilt",
 	"frame/grandpa",
@@ -120,12 +119,13 @@ members = [
 	"frame/system/benchmarking",
 	"frame/system/rpc/runtime-api",
 	"frame/timestamp",
+	"frame/tips",
 	"frame/transaction-payment",
 	"frame/transaction-payment/rpc",
 	"frame/transaction-payment/rpc/runtime-api",
 	"frame/transaction-storage",
 	"frame/treasury",
-	"frame/tips",
+	"frame/try-runtime",
 	"frame/uniques",
 	"frame/utility",
 	"frame/vesting",
@@ -161,15 +161,16 @@ members = [
 	"primitives/offchain",
 	"primitives/panic-handler",
 	"primitives/rpc",
-	"primitives/runtime",
 	"primitives/runtime-interface",
 	"primitives/runtime-interface/proc-macro",
-	"primitives/runtime-interface/test",
-	"primitives/runtime-interface/test-wasm",
 	"primitives/runtime-interface/test-wasm-deprecated",
+	"primitives/runtime-interface/test-wasm",
+	"primitives/runtime-interface/test",
+	"primitives/runtime",
 	"primitives/sandbox",
 	"primitives/serializer",
 	"primitives/session",
+	"primitives/ss58",
 	"primitives/staking",
 	"primitives/state-machine",
 	"primitives/std",
@@ -190,17 +191,17 @@ members = [
 	"test-utils/runtime",
 	"test-utils/runtime/client",
 	"test-utils/runtime/transaction-pool",
-	"test-utils/test-runner",
 	"test-utils/test-crate",
+	"test-utils/test-runner",
 	"utils/browser",
 	"utils/build-script-utils",
 	"utils/fork-tree",
 	"utils/frame/benchmarking-cli",
-	"utils/frame/remote-externalities",
 	"utils/frame/frame-utilities-cli",
-	"utils/frame/try-runtime/cli",
+	"utils/frame/remote-externalities",
 	"utils/frame/rpc/support",
 	"utils/frame/rpc/system",
+	"utils/frame/try-runtime/cli",
 	"utils/prometheus",
 	"utils/wasm-builder",
 ]
