+++ primitives/ss58/README.md
@@ -1,0 +1,3 @@
+Groups all common Substrate's Base58 logic.
+
+License: Apache-2.0
